libstore/derivation: Split hash modulo out of the ATerm format
`derivation/aterm.{hh,cc}` was carrying two topics: the ATerm encoding
of a derivation, and the hash-modulo computation defined over it. Now,
move the latter to `derivation/modulo.{hh,cc}` and put it in a new
`nix::derivation::modulo` (sub) namespace, changing identifiers
accordingly.

When I split out those files in 7d4eb33c21f421d9c25ec2636250722dd063a1ed
I thought putting both things in there was fine, but now I no longer think
it is, and we need to separate concerns even more. The unit tests really
want it.

The one thing the two files must share is the unparser for the
intermediate form, so the formerly-`static` `unparseDerivation` becomes
a public `unparse` overload in `aterm.hh`, explicitly instantiated in
`aterm.cc` for the `modulo::Inputs` cases. `aterm.hh` only forward
declares `modulo::Inputs`, so it does not depend on `modulo.hh`.

`src/libstore-tests/derivation/aterm.cc` moves to `derivation/modulo.cc`
to match, and into the new namespace with it; its golden masters already
lived in `data/derivation/hash-modulo`.

No behaviour changes. Recommended to review locally with
`git show --color-moved`.
