+++ src/libexpr/primops.cc
@@ -1,5 +1,5 @@
 #include "nix/store/derivations.hh"
-#include "nix/store/derivation/aterm.hh"
+#include "nix/store/derivation/modulo.hh"
 #include "nix/store/downstream-placeholder.hh"
 #include "nix/expr/eval-inline.hh"
 #include "nix/expr/eval.hh"
@@ -1926,8 +1926,8 @@ static void derivationStrictInternal(EvalState & state, std::string_view drvName
        case we don't actually write store derivations, so we can't
        read them later. */
     {
-        auto h = hashInputModulo(*state.store, drv);
-        derivation::hashes.insert_or_assign(drvPath, std::move(h));
+        auto h = derivation::modulo::hashInput(*state.store, drv);
+        derivation::modulo::hashes.insert_or_assign(drvPath, std::move(h));
     }
 
     auto result = state.buildBindings(1 + drv.outputs.size());
