+++ src/libstore-tests/meson.build
@@ -37,10 +37,10 @@ sources = files(
   'common-protocol.cc',
   'content-address.cc',
   'derivation/advanced-attrs.cc',
-  'derivation/aterm.cc',
   'derivation/external-formats.cc',
   'derivation/full-inputs.cc',
   'derivation/invariants.cc',
+  'derivation/modulo.cc',
   'derivation/resolution.cc',
   'derivation/write.cc',
   'derived-path.cc',
